+++ python/ray/tests/test_advanced_4.py
@@ -6,7 +6,12 @@ import pytest
 
 import ray
 from ray._private.gcs_utils import check_health
-from ray._private.test_utils import Semaphore, client_test_enabled, wait_for_condition
+from ray._private.test_utils import (
+    Semaphore,
+    client_test_enabled,
+    wait_for_condition,
+    get_gcs_memory_used,
+)
 from ray.experimental.internal_kv import _internal_kv_list
 
 
@@ -180,24 +185,6 @@ def test_local_mode_deadlock(shutdown_only_with_initialization_check):
     bar = Bar.remote()
     # Expect ping_actor call returns normally without deadlock.
     assert ray.get(foo.ping_actor.remote(bar)) == 3
-
-
-def get_gcs_memory_used():
-    import psutil
-
-    m = 0
-    for process in psutil.process_iter():
-        try:
-            if process.name() in ("gcs_server", "redis-server"):
-                m += process.memory_info().rss
-        except psutil.NoSuchProcess:
-            # NOTE(rickyx):
-            # From https://psutil.readthedocs.io/en/latest/#psutil.process_iter
-            # If you’re not interested in retrieving zombies just
-            # catch NoSuchProcess.
-            continue
-
-    return m
 
 
 def function_entry_num(job_id):
